travis: Make macOS' CI use Xcode 10
Keeps the toolchain used for CI up to date. Xcode 10 also brings in
better (but not complete) C++17 support on macOS. Notably it allows the
use of the <any>, <optional>, and <variant> headers that were introduced
in C++17, among other things. Now we can use the standard-provided
facilities instead of needing to rely on boost for these.